Move XmlToJsonConverter to jsonUtils module
